pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "src"
version = "0.1.0"
description = "Pipeline ETL para dados da ANM"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Vinicius Kuhlmann", email = "vinikuhlmann@gmail.com" }]
classifiers = [
    "Programming Language :: Python :: 3",
    # "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "python-dotenv",
    "certifi",         # Certificados SSL
    "requests",
    "pandas",
    "dask[dataframe]",
    "psycopg2-binary",
    "sqlalchemy",
]

[tool.setuptools]
packages = ["src"]
